+++ android/quest/src/test/java/org/smartregister/fhircore/quest/data/PatientRepositoryTest.kt
@@ -22,6 +22,9 @@ import dagger.hilt.android.testing.HiltAndroidTest
 import io.mockk.coEvery
 import io.mockk.coVerify
 import io.mockk.mockk
+import java.time.LocalDate
+import java.time.ZoneId
+import java.util.Date
 import javax.inject.Inject
 import kotlinx.coroutines.test.runBlockingTest
 import org.hl7.fhir.r4.model.Coding
@@ -96,15 +99,29 @@ class PatientRepositoryTest : RobolectricTest() {
   @Test
   fun testFetchTestResultsShouldReturnListOfTestReports() =
     coroutineTestRule.runBlockingTest {
+      val today = Date.from(LocalDate.now().atStartOfDay(ZoneId.systemDefault()).toInstant())
+      val yesterday =
+        Date.from(LocalDate.now().minusDays(1).atStartOfDay(ZoneId.systemDefault()).toInstant())
+
       coEvery { fhirEngine.search<QuestionnaireResponse>(any()) } returns
         listOf(
           QuestionnaireResponse().apply {
+            meta = Meta().apply { tag = listOf(Coding().apply { display = "Cell Count" }) }
+            authored = today
+          },
+          QuestionnaireResponse().apply {
             meta = Meta().apply { tag = listOf(Coding().apply { display = "Blood Count" }) }
+            authored = yesterday
           }
         )
 
       val results = repository.fetchTestResults("1")
-      Assert.assertEquals("Blood Count", results.first().meta?.tagFirstRep?.display)
+
+      Assert.assertEquals("Cell Count", results.first().meta?.tagFirstRep?.display)
+      Assert.assertEquals(today.time, results.first().authored?.time)
+
+      Assert.assertEquals("Blood Count", results.last().meta?.tagFirstRep?.display)
+      Assert.assertEquals(yesterday.time, results.last().authored?.time)
     }
 
   @Test