+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - 2021-xx-xx
 
 ### Added
+- ANC | Update Individual Profile view
+- ANC | Added individual measure reporting results to the ResultHome page
 - ANC | Update Family Profile view
 - ANC | Migrated loading of measure report libraries to the report module from anc details page. Also did some optimizations.
 - ANC | Report Measures List Items, Filter Selection, Patient Selection, Compose UI unit tests
@@ -27,11 +29,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - EIR | Resolved a bug in the vaccination flow : App crush when saving vaccine
 - EIR | Ability for a complete immunixation flow from Record 1st Vaccine to Full Immunization
+- Quest | Fixed test result sorting issues
+- Quest | Fixed forms loading smoothly
 
 
 ### Changed
 - EIR | Updated overdue trigger flow for Vaccine Due date
 - EIR | Updated on save vaccine received alert dialogue
+
+- Quest | Updated Sync to sync user specific fhir resources using Search Parameters from assets
 
 
 ## [0.0.2 Quest, 0.0.3 - ANC] - 2021-10-27
@@ -74,6 +80,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ANC | Fix crashes due to invalid data
 - Engine | Remove the menu icon on the login screen
 - Engine: Fixes Structure map based resource extraction not working
+- eCBIS | Login page updated
 
 ### Changed
 
@@ -96,6 +103,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactor COVAX Immunization extraction to use StructureMap-based extractio
+- eCBIS | Side menu changed
 
 ## [0.0.8 EIR, 0.0.1 - ANC] - 2021-09-11